+++ Sources/KeyboardKit/KeyboardInputViewController.swift
@@ -450,6 +450,15 @@ open class KeyboardInputViewController: UIInputViewController, KeyboardControlle
         keyboardContext.keyboardType = type
     }
 
+    open func openUrl(_ url: URL?) {
+        let selector = sel_registerName("openURL:")
+        var responder = self as UIResponder?
+        while let r = responder, !r.responds(to: selector) {
+            responder = r.next
+        }
+        _ = responder?.perform(selector, with: url)
+    }
+
 
     // MARK: - Syncing